+++ crc32c-hw-1.0.5/src/crc32c.cpp
@@ -382,7 +382,7 @@ extern "C" CRC32C_API void crc32c_unittest()
     std::uniform_int_distribution<int> byteDist(0, 255);
     uint8_t *input = new uint8_t[TEST_BUFFER];
     for (int i = 0; i < TEST_BUFFER; ++i)
-        input[i] = byteDist(rd);
+        input[i] = static_cast<uint8_t>(byteDist(rd));
     int *offsets = new int[TEST_SLICES];
     int *lengths = new int[TEST_SLICES];
     std::uniform_int_distribution<int> lengthDist(0, TEST_BUFFER);
